# openapi.yaml
openapi: 3.0.3
info:
  title: Young Artist Community API
  description: |
    REST API for Young Artist Community, a directory and review platform for
    Young Artist Programs (YAPs) in classical music and opera.
  version: 0.2.0
  contact:
    name: Young Artist Community
    email: api@youngartist.community
    url: https://youngartist.community
  x-api-id: young-artist-community-api
  x-audience: external-public
servers:
  - url: https://youngartist.community

# ---------------------------------------------------------------------------
# PATHS
# ---------------------------------------------------------------------------
paths:
...
  /programs:
    get:
      summary: List and filter programs
      description: |
        Returns a cursor-paginated list of programs. Supports filtering by
        instrument, category, location, and deadline status. Supports sorting
        via the `sort` query parameter.
      operationId: listPrograms
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Sort'
        - name: instrument_id
          in: query
          description: Filter by instrument UUID. Supports comma-separated values for OR semantics.
          schema:
            type: string
        - name: category_id
          in: query
          description: Filter by category UUID. Supports comma-separated values for OR semantics.
          schema:
            type: string
        - name: location_id
          in: query
          description: Filter by location UUID.
          schema:
            type: string
        - name: country
          in: query
          description: Filter by country name (exact match).
          schema:
            type: string
        - name: deadline_after
          in: query
          description: Return programs with application_deadline after this date (ISO 8601).
          schema:
            type: string
            format: date
        - name: deadline_before
          in: query
          description: Return programs with application_deadline before this date (ISO 8601).
          schema:
            type: string
            format: date
        - name: tuition_lower_than
          in: query
          description: Maximum tuition (inclusive upper bound).
          schema:
            type: number
            format: float
        - name: offers_scholarship
          in: query
          description: Filter to programs that offer scholarships.
          schema:
            type: boolean
        - name: q
          in: query
          description: Free-text search across program name and description.
          schema:
            type: string
      responses:
        '200':
          description: A cursor-paginated list of programs.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProgramListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
        default:
          $ref: '#/components/responses/DefaultError'

    post:
      summary: Create a program
      description: Creates a new program listing. Intended for admin or future authenticated use.
      operationId: createProgram
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProgramInput'
      responses:
        '201':
          description: Program created successfully.
          headers:
            Location:
              description: URL of the newly created program.
              schema:
                type: string
                format: uri
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Program'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
        default:
          $ref: '#/components/responses/DefaultError'
  /programs/{program_id}:
    get:
      summary: Get a single program
      operationId: getProgram
      parameters:
        - $ref: '#/components/parameters/ProgramId'
      responses:
        '200':
          description: The requested program including computed average_rating.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Program'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
        default:
          $ref: '#/components/responses/DefaultError'

    put:
      summary: Update a program
      description: Replaces the entire program resource.
      operationId: updateProgram
      parameters:
        - $ref: '#/components/parameters/ProgramId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProgramInput'
      responses:
        '200':
          description: Program updated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Program'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
        default:
          $ref: '#/components/responses/DefaultError'

  # ---- Reviews (sub-resource of Program) ----------------------------------
...
components:
...
  parameters:
    ProgramId:
      name: program_id
      in: path
      required: true
      description: UUID of the program.
      schema:
        type: string
        format: uuid
...
    Cursor:
      name: cursor
      in: query
      description: |
        Opaque cursor token identifying the start of the page. Obtain from
        the `next` or `prev` field of a previous list response. Omit to
        fetch the first page.
      schema:
        type: string
    Limit:
      name: limit
      in: query
      description: Maximum number of items per page.
      schema:
        type: integer
        format: int32
        minimum: 1
        maximum: 100
        default: 20
    Sort:
      name: sort
      in: query
      description: |
        Comma-separated list of fields to sort by. Prefix with `-` for
        descending order. Example: `-created_at,name`
      schema:
        type: string

  # ---- Reusable responses -------------------------------------------------
  responses:
    BadRequest:
      description: Malformed request or invalid parameters.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    NotFound:
      description: The requested resource was not found.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    InternalServerError:
      description: Unexpected server error.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    DefaultError:
      description: Unexpected error.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'

  # ---- Schemas ------------------------------------------------------------
  schemas:
...
    Problem:
      type: object
      description: Problem detail object per RFC 9457.
      properties:
        type:
          type: string
          format: uri-reference
          default: about:blank
          description: |
            A URI reference that identifies the problem type. When dereferenced,
            it SHOULD provide human-readable documentation for the problem type.
          example: /problems/not-found
        title:
          type: string
          description: A short human-readable summary of the problem type.
          example: Resource not found
        status:
          type: integer
          format: int32
          minimum: 100
          maximum: 600
          exclusiveMaximum: true
          description: The HTTP status code generated for this occurrence.
          example: 404
        detail:
          type: string
          description: Human-readable explanation specific to this occurrence.
          example: Program with id '550e8400-e29b-41d4-a716-446655440000' not found.
        instance:
          type: string
          format: uri-reference
          description: A URI reference that identifies this specific occurrence of the problem.

    # -- Cursor pagination meta --
    CursorPageMeta:
      type: object
      description: |
        Cursor-based pagination metadata. `next` and `prev` are opaque tokens
        passed to the `cursor` query parameter on subsequent requests. Both
        are null when no further page exists in that direction.
      properties:
        next:
          type: string
          nullable: true
          description: Opaque cursor for the next page, or null if this is the last page.
        prev:
          type: string
          nullable: true
          description: Opaque cursor for the previous page, or null if this is the first page.
        total_items:
          type: integer
          format: int32
          description: Total number of items matching the query across all pages.

    # -- Program --
    Program:
      type: object
      required:
        - id
        - slug
        - name
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        slug:
          type: string
          readOnly: true
          description: URL-friendly kebab-case identifier derived from the program name.
          example: salzburg-festival-young-singers-project
        name:
          type: string
          example: Salzburg Festival Young Singers Project
        description:
          type: string
        start_date:
          type: string
          format: date
        end_date:
          type: string
          format: date
        application_deadline:
          type: string
          format: date
        tuition:
          type: number
          format: float
          example: 2500.00
        application_fee:
          type: number
          format: float
          example: 75.00
        age_min:
          type: integer
          format: int32
        age_max:
          type: integer
          format: int32
        offers_scholarship:
          type: boolean
          default: false
        application_url:
          type: string
          format: uri
        program_url:
          type: string
          format: uri
        average_rating:
          type: number
          format: float
          readOnly: true
          description: Computed average of all review ratings (1.0 to 5.0).
        review_count:
          type: integer
          format: int32
          readOnly: true
        instruments:
          type: array
          readOnly: true
          items:
            $ref: '#/components/schemas/Instrument'
        categories:
          type: array
          readOnly: true
          items:
            $ref: '#/components/schemas/Category'
        locations:
          type: array
          readOnly: true
          items:
            $ref: '#/components/schemas/Location'
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
    ProgramInput:
      type: object
      required:
        - name
      properties:
        name:
          type: string
        description:
          type: string
        start_date:
          type: string
          format: date
        end_date:
          type: string
          format: date
        application_deadline:
          type: string
          format: date
        tuition:
          type: number
          format: float
        application_fee:
          type: number
          format: float
        age_min:
          type: integer
          format: int32
        age_max:
          type: integer
          format: int32
        offers_scholarship:
          type: boolean
        application_url:
          type: string
          format: uri
        program_url:
          type: string
          format: uri
        instrument_ids:
          type: array
          items:
            type: string
            format: uuid
        category_ids:
          type: array
          items:
            type: string
            format: uuid
        location_ids:
          type: array
          items:
            type: string
            format: uuid
    ProgramListResponse:
      type: object
      required:
        - items
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Program'
        meta:
          $ref: '#/components/schemas/CursorPageMeta'

    # -- Review --
...
    Instrument:
      type: object
      required:
        - id
        - name
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: violin
    Category:
      type: object
      required:
        - id
        - name
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: opera
    Location:
      type: object
      required:
        - id
        - city
        - country
      properties:
        id:
          type: string
          format: uuid
        city:
          type: string
          example: Salzburg
        country:
          type: string
          example: Austria
        state:
          type: string
        address:
          type: string